circuit stacking bug fix using deep copies
Easiest solution was to make a deep copy of all the circuits involved.  This solved the reindexing problem, which was also the root of the qubit_indices length error that occurred when you tried to stack more than 2 circuits which referenced the same object.

Also found a relic "agnostic_simulator" reference in ADAPTSolver description